+++ test/utils/node_map.cpp
@@ -5,11 +5,12 @@
 
 #include <mockturtle/generators/arithmetic.hpp>
 #include <mockturtle/networks/aig.hpp>
+#include <mockturtle/networks/mig.hpp>
 #include <mockturtle/utils/node_map.hpp>
 
 using namespace mockturtle;
 
-TEST_CASE( "create node map for full adder", "[node_map]" )
+TEST_CASE( "create vector node map for full adder", "[node_map]" )
 {
   aig_network aig;
 
@@ -42,4 +43,54 @@ TEST_CASE( "create node map for full adder", "[node_map]" )
   } );
 
   CHECK( total == aig.size() );
+}
+
+TEST_CASE( "create unordered node map for full adder", "[node_map]" )
+{
+  mig_network mig;
+
+  const auto a = mig.create_pi();
+  const auto b = mig.create_pi();
+  const auto c = mig.create_pi();
+
+  const auto [sum, carry] = full_adder( mig, a, b, c );
+
+  mig.create_po( sum );
+  mig.create_po( carry );
+
+  unordered_node_map<uint32_t, mig_network> map( mig );
+  mig.foreach_node( [&]( auto n ) {
+    CHECK( !map.has( n ) );
+  } );
+
+  mig.foreach_node( [&]( auto n, auto i ) {
+    map[n] = i;
+  } );
+
+  mig.foreach_node( [&]( auto n ) {
+    CHECK( map.has( n ) );
+  } );
+
+  uint32_t total{0};
+  mig.foreach_node( [&]( auto n ) {
+    total += map[n];
+  } );
+
+  CHECK( total == ( mig.size() * ( mig.size() - 1 ) ) / 2 );
+
+  map.reset();
+  mig.foreach_node( [&]( auto n ) {
+    CHECK( !map.has( n ) );
+  } );
+
+  mig.foreach_node( [&]( auto n ) {
+    map[n] = 1;
+  } );
+
+  total = 0;
+  mig.foreach_node( [&]( auto n ) {
+    total += map[n];
+  } );
+
+  CHECK( total == mig.size() );
 }